+++ src/polyline.rs
@@ -572,10 +572,10 @@ impl Polyline {
             linecolor: Some(crate::proto::Color {
                 guid: self.linecolor.guid().to_string(),
                 name: self.linecolor.name.clone(),
-                r: self.linecolor.r as i32,
-                g: self.linecolor.g as i32,
-                b: self.linecolor.b as i32,
-                a: self.linecolor.a as i32,
+                r: self.linecolor.r,
+                g: self.linecolor.g,
+                b: self.linecolor.b,
+                a: self.linecolor.a,
             }),
             xform: Some(crate::proto::Xform {
                 guid: self.xform.guid().to_string(),
@@ -608,10 +608,10 @@ impl Polyline {
         if let Some(color) = proto.linecolor {
             pl.linecolor.set_guid(color.guid);
             pl.linecolor.name = color.name;
-            pl.linecolor.r = color.r as u8;
-            pl.linecolor.g = color.g as u8;
-            pl.linecolor.b = color.b as u8;
-            pl.linecolor.a = color.a as u8;
+            pl.linecolor.r = color.r;
+            pl.linecolor.g = color.g;
+            pl.linecolor.b = color.b;
+            pl.linecolor.a = color.a;
         }
 
         if let Some(xform) = proto.xform {
